+++ src/pykeen/nn/sim.py
@@ -7,7 +7,7 @@ import torch
 from class_resolver import ClassResolver
 from torch import nn
 
-from .compute_kernel import batched_dot
+from .functional import batched_dot
 from ..typing import FloatTensor, GaussianDistribution
 from ..utils import at_least_eps, tensor_sum
