+++ src/pykeen/nn/text/__init__.py
@@ -1,0 +1,18 @@
+"""Utilities for text lookup, caching, and encoding."""
+
+from .text_cache import IdentityCache, PyOBOCache, TextCache, WikidataTextCache, text_cache_resolver
+from .text_encoder import CharacterEmbeddingTextEncoder, TextEncoder, TransformerTextEncoder, text_encoder_resolver
+
+__all__ = [
+    # Text Cache
+    "text_cache_resolver",
+    "TextCache",
+    "IdentityCache",
+    "PyOBOCache",
+    "WikidataTextCache",
+    # Text Encoder
+    "text_encoder_resolver",
+    "TextEncoder",
+    "CharacterEmbeddingTextEncoder",
+    "TransformerTextEncoder",
+]
